+++ data/data_manager.rb
@@ -1,0 +1,4 @@
+require 'json'
+
+class DataManager
+end
