+++ lib/app.rb
