+++ lib/book.rb
