+++ lib/classroom.rb
