+++ lib/display.rb
