+++ lib/display_option.rb
