+++ lib/nameable.rb
