+++ lib/rental.rb
