+++ lib/student.rb
@@ -3,7 +3,7 @@ require_relative 'person'
 class Student < Person
   attr_reader :classroom
 
-  def initialize(age, classroom, parent_permission, name = 'unknow')
+  def initialize(age, classroom, name, parent_permission)
     super(age, name, parent_permission: parent_permission)
     @classroom = classroom
   end
