+++ lib/teacher.rb
@@ -3,7 +3,7 @@ require_relative 'person'
 class Teacher < Person
   attr_reader :specialization
 
-  def initialize(age, specialization, name = 'unknow')
+  def initialize(age, name, specialization)
     super(age, name)
     @specialization = specialization
   end
